+++ JUN_All/tools/A00275_skinTool_V01/app/core/__init__.py
@@ -1,4 +1,5 @@
 from .skin_migrate_manager import SkinMigrateManager
 from . import bind_pose_manager
+from . import weight_transfer_manager
 
-__all__ = ["SkinMigrateManager", "bind_pose_manager"]
+__all__ = ["SkinMigrateManager", "bind_pose_manager", "weight_transfer_manager"]